refactor sf e2e actions
